requirements.txt
diff-match-patch==20121119
Django==2.0.6
django-datatable==0.3.1
django-datatables-view==1.16.0
-e git+https://github.com/mjs7231/django-dbbackup.git@c71b292d68759322c120629bd71352da30406f04#egg=django_dbbackup
django-import-export==1.0.1
django-tables2==1.21.2
djangorestframework==3.8.2
djangorestframework-datatables==0.4.0
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.4
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
django-grappelli==2.11.1